Move keystone stuff out of plugins folder
