+++ keystone/models/gallery-items.js
