+++ lib/routes.js
@@ -12,7 +12,7 @@ const series = require('./controllers/series');
 const config = require('./config');
 const indexing = require('../controllers/indexing');
 const users = require('../controllers/users');
-const keystonePlugin = require('../plugins/keystone');
+const keystonePlugin = require('../keystone');
 const authPlugin = require('../auth');
 
 /**
